+SetTextureEx exported to lua for CUIStatic, which allows setting shader name as second param = UIDragDropListEx | Fix improper use of std::unique. The idea is to cull cells that are the same. ~ uncommented multithreading console commands

+++ src/xrGame/ai/monsters/rats/ai_rat.cpp
@@ -487,7 +487,7 @@ void CAI_Rat::UpdateCL			()
 
 		if (squad && 
 			((squad->GetLeader() != this && !squad->GetLeader()->g_Alive()) ||
-			squad->get_index(this) == u32(-1))
+			squad->get_index(this) == u8(-1))
 			) squad->SetLeader(this);
 
 		if (squad &&
